CODE THIEN

+++ attendance.py
@@ -10,7 +10,7 @@ import shutil
 from tkinter import filedialog
 import os ,string ,time
 from send_gmail import Attach
-
+from time import strftime
 mydata=[]
 
 
@@ -18,10 +18,7 @@ def diff(list1, list2):
     list_difference = [item for item in list1 if item not in list2]
     return list_difference
 
-# def send_gmail():
-#     win=Tk()
-#     app=Attendance(win)
-#     win.mainloop()
+
 
 class Attendance:
     def __init__(self,root):
@@ -31,7 +28,6 @@ class Attendance:
 
         #=================variables=========================
         self.var_atten_id=StringVar()
-        self.var_atten_roll = StringVar()
         self.var_atten_name = StringVar()
         self.var_atten_dep = StringVar()
         self.var_atten_time = StringVar()
@@ -40,7 +36,7 @@ class Attendance:
 
 
         #first image
-        img=Image.open(r"college_images\logocaothang.png")
+        img=Image.open(r"college_images\Cap3.PNG")
         img=img.resize((800,200), Image.ANTIALIAS)
         self.photoimg=ImageTk.PhotoImage(img)
 
@@ -48,7 +44,7 @@ class Attendance:
         f_lbl.place(x=0, y=0, width=800, height=200)
 
         #second image
-        img1=Image.open(r"college_images\class.jpg")
+        img1=Image.open(r"college_images\Cap2.PNG")
         img1=img1.resize((800,200), Image.ANTIALIAS)
         self.photoimg1=ImageTk.PhotoImage(img1)
 
@@ -67,6 +63,15 @@ class Attendance:
         title_lbl=Label(bg_img,text="ATTENDANCE MANAGEMENT SYSTEM",font=("times new roman",35,"bold"),bg="white",fg="darkgreen")
         title_lbl.place(x=0,y=0,width=1530, height=45)
 
+
+        def time():
+            string= strftime('%H:%M:%S %p')
+            lbl.config(text = string)
+            lbl.after(1000,time)
+
+        lbl = Label(title_lbl, font=('times new roman',14,'bold'),background='white',foreground='blue')
+        lbl.place(x=0,y=0,width=110, height=50)
+        time()
 
         main_frame=Frame(bg_img, bd=2, bg="white")
         main_frame.place(x=20, y=50, width=1480, height=600)
@@ -87,22 +92,22 @@ class Attendance:
 
         #Labeland entry
         #attendance id
-        attendance_label=Label(left_inside_frame,text="AttendanceId", font=("times new roman",13,"bold"),bg="white")
+        attendance_label=Label(left_inside_frame,text="Mã Số Sinh Viên:", font=("times new roman",13,"bold"),bg="white")
         attendance_label.grid(row=0,column=0,padx=10,pady=5,sticky=W)
 
         attendance_entry=ttk.Entry(left_inside_frame,width=20,textvariable=self.var_atten_id,font=("times new roman",13,"bold"))
         attendance_entry.grid(row=0,column=1,padx=10,pady=5,sticky=W)
 
-        #Roll
-        rollLabel=Label(left_inside_frame,text="Roll:",bg="white", font="comicsansns 11 bold")
-        rollLabel.grid(row=0,column=2,padx=4,pady=8)
-
-        atten_roll=ttk.Entry(left_inside_frame,width=22,textvariable=self.var_atten_roll,font="comicsansns 11 bold")
-        atten_roll.grid(row=0,column=3,pady=8)
+        # #Roll
+        # rollLabel=Label(left_inside_frame,text="Roll:",bg="white", font="comicsansns 11 bold")
+        # rollLabel.grid(row=0,column=2,padx=4,pady=8)
+        #
+        # atten_roll=ttk.Entry(left_inside_frame,width=22,textvariable=self.var_atten_roll,font="comicsansns 11 bold")
+        # atten_roll.grid(row=0,column=3,pady=8)
 
 
         #Name
-        nameLabel=Label(left_inside_frame,text="Name:",bg="white", font="comicsansns 11 bold")
+        nameLabel=Label(left_inside_frame,text="Tên Sinh Viên:",bg="white", font="comicsansns 11 bold")
         nameLabel.grid(row=1,column=0)
 
         atten_name=ttk.Entry(left_inside_frame,width=22,textvariable=self.var_atten_name,font="comicsansns 11 bold")
@@ -110,36 +115,33 @@ class Attendance:
 
 
         #Department
-        depLabel=Label(left_inside_frame,text="Department:",bg="white", font="comicsansns 11 bold")
-        depLabel.grid(row=1,column=2)
+        depLabel=Label(left_inside_frame,text="Môn Học:",bg="white", font="comicsansns 11 bold")
+        depLabel.grid(row=0,column=2)
 
         atten_dep=ttk.Entry(left_inside_frame,width=22,textvariable=self.var_atten_dep,font="comicsansns 11 bold")
-        atten_dep.grid(row=1,column=3,pady=8)
+        atten_dep.grid(row=0,column=3,pady=8)
 
         #Time
-        timeLabel=Label(left_inside_frame,text="Time:",bg="white", font="comicsansns 11 bold")
+        timeLabel=Label(left_inside_frame,text="Thời Gian Có Mặt:",bg="white", font="comicsansns 11 bold")
         timeLabel.grid(row=2,column=0)
 
         atten_time=ttk.Entry(left_inside_frame,width=22,textvariable=self.var_atten_time,font="comicsansns 11 bold")
         atten_time.grid(row=2,column=1,pady=8)
 
         #Date
-        dateLabel=Label(left_inside_frame,text="Date:",bg="white", font="comicsansns 11 bold")
-        dateLabel.grid(row=2,column=2)
+        dateLabel=Label(left_inside_frame,text="Ngày Tham Dự:",bg="white", font="comicsansns 11 bold")
+        dateLabel.grid(row=1,column=2)
 
         atten_date=ttk.Entry(left_inside_frame,width=22,textvariable=self.var_atten_date,font="comicsansns 11 bold")
-        atten_date.grid(row=2,column=3,pady=8)
+        atten_date.grid(row=1,column=3,pady=8)
 
 
         #attendance
-        attendanceLabel=Label(left_inside_frame,text="Attendance Status:",bg="white", font="comicsansns 11 bold")
-        attendanceLabel.grid(row=3,column=0)
+        attendanceLabel=Label(left_inside_frame,text="Tham Dự:",bg="white", font="comicsansns 11 bold")
+        attendanceLabel.grid(row=2,column=2)
 
-        self.atten_status=ttk.Combobox(left_inside_frame,width=20,textvariable=self.var_atten_attendance,font="comicsansns 11 bold",state="readonly")
-        self.atten_status["values"]=("Status","Present","Absent")
-        self.atten_status.grid(row=3,column=1,pady=8)
-        self.atten_status.current(0)
-
+        atten_status=ttk.Entry(left_inside_frame,width=22,textvariable=self.var_atten_attendance,font="comicsansns 11 bold")
+        atten_status.grid(row=2,column=3,pady=8)
 
         #buttons frame
         btn_frame=Frame(left_inside_frame, bd=2, relief=RIDGE, bg="white")
@@ -182,7 +184,7 @@ class Attendance:
         scroll_x=ttk.Scrollbar(table_frame, orient=HORIZONTAL)
         scroll_y=ttk.Scrollbar(table_frame, orient=VERTICAL)
 
-        self.AttendanceReportTable=ttk.Treeview(table_frame, columns=("id","roll","name","department","time","date","attendance"),xscrollcommand=scroll_x.set,yscrollcommand=scroll_y.set)
+        self.AttendanceReportTable=ttk.Treeview(table_frame, columns=("id","name","department","time","date","attendance"),xscrollcommand=scroll_x.set,yscrollcommand=scroll_y.set)
 
         scroll_x.pack(side=BOTTOM,fill=X)
         scroll_y.pack(side=RIGHT, fill=Y)
@@ -190,18 +192,16 @@ class Attendance:
         scroll_x.config(command=self.AttendanceReportTable.xview)
         scroll_y.config(command=self.AttendanceReportTable.yview)
 
-        self.AttendanceReportTable.heading("id",text="Attendane")
-        self.AttendanceReportTable.heading("roll", text="StudentID")
-        self.AttendanceReportTable.heading("name", text="Name")
-        self.AttendanceReportTable.heading("department", text="Department")
-        self.AttendanceReportTable.heading("time", text="Time")
-        self.AttendanceReportTable.heading("date", text="Date")
-        self.AttendanceReportTable.heading("attendance", text="Attendance")
+        self.AttendanceReportTable.heading("id",text="MSSV")
+        self.AttendanceReportTable.heading("name", text="Tên")
+        self.AttendanceReportTable.heading("department", text="Môn Học")
+        self.AttendanceReportTable.heading("time", text="Thời Gian")
+        self.AttendanceReportTable.heading("date", text="Ngày")
+        self.AttendanceReportTable.heading("attendance", text="Hiện Diện")
 
         self.AttendanceReportTable["show"]="headings"
 
         self.AttendanceReportTable.column("id",width=100)
-        self.AttendanceReportTable.column("roll", width=100)
         self.AttendanceReportTable.column("name", width=150)
         self.AttendanceReportTable.column("department", width=100)
         self.AttendanceReportTable.column("time", width=100)
@@ -228,24 +228,24 @@ class Attendance:
         global mydata
         mydata.clear()
         fln=filedialog.askopenfilename(initialdir=os.getcwd(),title="Open CSV",filetypes=(("CSV File","*.csv"),("ALL File","*.*")),parent=self.root)
-        with open(fln) as myfile:
+        with open(fln) as myfile:  #mở tệp  csv
             csvread=csv.reader(myfile, delimiter=",")
             for i in csvread:
                 mydata.append(i)
             self.fetchData(mydata)
 
-    #export csv
+
     def exportCsv(self):
         try:
             if len(mydata)<1:
                 messagebox.showerror("No Data","No Data Found to Export",parent=self.root)
                 return False
-            fln = filedialog.asksaveasfilename(initialdir=os.getcwd(), title="Open CSV",filetypes=(("CSV File", "*.csv"), ("ALL File", "*.*")), parent=self.root)
+            fln = filedialog.asksaveasfilename(initialdir=os.getcwd(), title="Export CSV",filetypes=(("CSV File", "*.csv"), ("ALL File", "*.*")), parent=self.root)
             with open(fln, mode="w",newline="") as myfile:
                 exp_write=csv.writer(myfile, delimiter=",")
                 for i in mydata:
                     exp_write.writerow(i)
-                messagebox.showinfo("Data Export","Your data exported to"+os.path.basename(fln)+"successfully")
+                messagebox.showinfo("Data Export","Your data exported to "+os.path.basename(fln)+" successfully",parent=self.root)
         except Exception as es:
             messagebox.showerror("Error", f"Due to :{str(es)}", parent=self.root)
 
@@ -254,12 +254,11 @@ class Attendance:
         content=self.AttendanceReportTable.item(cursor_row)
         rows=content['values']
         self.var_atten_id.set(rows[0])
-        self.var_atten_roll.set(rows[1])
-        self.var_atten_name.set(rows[2])
-        self.var_atten_dep.set(rows[3])
-        self.var_atten_time.set(rows[4])
-        self.var_atten_date.set(rows[5])
-        self.var_atten_attendance.set(rows[6])
+        self.var_atten_name.set(rows[1])
+        self.var_atten_dep.set(rows[2])
+        self.var_atten_time.set(rows[3])
+        self.var_atten_date.set(rows[4])
+        self.var_atten_attendance.set(rows[5])
 
 
     #===Delete====
@@ -272,13 +271,8 @@ class Attendance:
                 global mydata
                 mydata.clear()
                 self.fetchData(mydata)
-                self.var_atten_id.set("")
-                self.var_atten_roll.set("")
-                self.var_atten_name.set("")
-                self.var_atten_dep.set("")
-                self.var_atten_time.set("")
-                self.var_atten_date.set("")
-                self.var_atten_attendance.set("")
+
+                self.reset_data()
                 messagebox.showinfo("Success","Student details successfully delete completed",parent=self.root)
                 open(r'C:\Users\PC_MINH THIEN\PycharmProjects\PythonProjectAI2\attend.csv', 'w')
         except Exception as es:
@@ -288,7 +282,6 @@ class Attendance:
     #===reset====
     def reset_data(self):
         self.var_atten_id.set("")
-        self.var_atten_roll.set("")
         self.var_atten_name.set("")
         self.var_atten_dep.set("")
         self.var_atten_time.set("")
@@ -301,24 +294,26 @@ class Attendance:
         dl = 'ABCDEFGHIJKLMNOPQRSTUVWXYZ'
         drives = ['%s:' % d for d in dl if os.path.exists('%s:' % d)]
         print(drives)
-        while True:
-            uncheckeddrives = ['%s:' % d for d in dl if os.path.exists('%s:' % d)]
-            x = diff(uncheckeddrives, drives)
-            if x:
-                x = "+".join(x)
-                filepath = (str(x) + "/Attend.csv")
-                print(filepath)  # E:/Attend.csv
-                print(os.path.exists(filepath))  # True or False
-                try:
-                    usb_file = messagebox.askyesno("Connect", "Connect USB successful", parent=self.root)
-                    if not (os.path.exists(filepath)):
-                        if (usb_file>0):
-                            open(filepath, 'w')
-                    shutil.copy(src=r'C:\Users\PC_MINH THIEN\PycharmProjects\PythonProjectAI2\attend.csv', dst=filepath)
-                    messagebox.showinfo("Success", "Exported to USB successfully", parent=self.root)
-                except Exception as es:
-                    messagebox.showerror("Error", f"Due to :{str(es)}", parent=self.root)
-                break
+        filesendusb = messagebox.askyesno("Please", "Please connect with your USB", parent=self.root)
+        if filesendusb >0:
+            while True:
+                uncheckeddrives = ['%s:' % d for d in dl if os.path.exists('%s:' % d)]
+                x = diff(uncheckeddrives, drives)
+                if x:
+                    x = "+".join(x)
+                    filepath = (str(x) + "/Attend.csv")
+                    print(filepath)
+                    print(os.path.exists(filepath))
+                    try:
+                        usb_file = messagebox.askyesno("Connect", "Do you want to send files to USB?", parent=self.root)
+                        if not (os.path.exists(filepath)):
+                            if (usb_file>0):
+                                open(filepath, 'w')
+                        shutil.copy(src=r'C:\Users\PC_MINH THIEN\PycharmProjects\PythonProjectAI2\attend.csv', dst=filepath)
+                        messagebox.showinfo("Success", "Exported to USB successfully", parent=self.root)
+                    except Exception as es:
+                        messagebox.showerror("Error", f"Due to :{str(es)}", parent=self.root)
+                    break
 
     def send_gmail(self):
         self.new_window=Toplevel(self.root)
